+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/command/gene/AddAllGenesCommand.kt
@@ -1,4 +1,4 @@
-package dev.aaronhowser.mods.geneticsresequenced.command
+package dev.aaronhowser.mods.geneticsresequenced.command.gene
 
 import com.mojang.brigadier.builder.ArgumentBuilder
 import com.mojang.brigadier.context.CommandContext
@@ -20,7 +20,7 @@ object AddAllGenesCommand {
 
     fun register(): ArgumentBuilder<CommandSourceStack, *> {
         return Commands
-            .literal("addAllGenes")
+            .literal("addAll")
             .requires { it.hasPermission(2) }
             .then(
                 Commands