+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/command/gene/AddGeneCommand.kt
@@ -1,11 +1,13 @@
-package dev.aaronhowser.mods.geneticsresequenced.command
+package dev.aaronhowser.mods.geneticsresequenced.command.gene
 
 import com.mojang.brigadier.arguments.StringArgumentType
 import com.mojang.brigadier.builder.ArgumentBuilder
 import com.mojang.brigadier.context.CommandContext
+import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.Gene
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.GeneRegistry
-import dev.aaronhowser.mods.geneticsresequenced.attachment.GenesData.Companion.removeGene
+import dev.aaronhowser.mods.geneticsresequenced.attachment.GenesData.Companion.addGene
+import dev.aaronhowser.mods.geneticsresequenced.attachment.GenesData.Companion.hasGene
 import dev.aaronhowser.mods.geneticsresequenced.command.ModCommands.SUGGEST_GENE_RLS
 import dev.aaronhowser.mods.geneticsresequenced.command.ModCommands.SUGGEST_GENE_STRINGS
 import dev.aaronhowser.mods.geneticsresequenced.datagen.ModLanguageProvider
@@ -19,7 +21,7 @@ import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.entity.Entity
 import net.minecraft.world.entity.LivingEntity
 
-object RemoveGeneCommand {
+object AddGeneCommand {
 
     private const val GENE_RL_ARGUMENT = "geneRl"
     private const val GENE_STRING_ARGUMENT = "geneString"
@@ -27,7 +29,7 @@ object RemoveGeneCommand {
 
     fun register(): ArgumentBuilder<CommandSourceStack, *> {
         return Commands
-            .literal("removeGene")
+            .literal("add")
             .requires { it.hasPermission(2) }
             .then(
                 Commands
@@ -39,7 +41,7 @@ object RemoveGeneCommand {
                             .then(
                                 Commands.argument(TARGET_ARGUMENT, EntityArgument.entities())
                                     .executes { cmd ->
-                                        removeGene(
+                                        addGene(
                                             cmd,
                                             StringArgumentType.getString(cmd, GENE_STRING_ARGUMENT),
                                             EntityArgument.getEntities(cmd, TARGET_ARGUMENT)
@@ -47,7 +49,7 @@ object RemoveGeneCommand {
                                     }
                             )
                             .executes { cmd ->
-                                removeGene(
+                                addGene(
                                     cmd,
                                     StringArgumentType.getString(cmd, GENE_STRING_ARGUMENT),
                                     entities = null
@@ -65,7 +67,7 @@ object RemoveGeneCommand {
                                 Commands
                                     .argument(TARGET_ARGUMENT, EntityArgument.entities())
                                     .executes { cmd ->
-                                        removeGene(
+                                        addGene(
                                             cmd,
                                             ResourceLocationArgument.getId(cmd, GENE_RL_ARGUMENT),
                                             EntityArgument.getEntities(cmd, TARGET_ARGUMENT)
@@ -73,7 +75,7 @@ object RemoveGeneCommand {
                                     }
                             )
                             .executes { cmd ->
-                                removeGene(
+                                addGene(
                                     cmd,
                                     ResourceLocationArgument.getId(cmd, GENE_RL_ARGUMENT),
                                     entities = null
@@ -84,19 +86,18 @@ object RemoveGeneCommand {
 
     }
 
-    private fun removeGene(
+    private fun addGene(
         context: CommandContext<CommandSourceStack>,
         geneRl: ResourceLocation,
         entities: MutableCollection<out Entity>? = null
     ): Int {
-
         val gene = GeneRegistry.fromResourceLocation(context.source.registryAccess(), geneRl)
             ?: throw IllegalArgumentException("Gene with id $geneRl does not exist!")
 
-        return removeGene(context, gene, entities)
+        return addGene(context, gene, entities)
     }
 
-    private fun removeGene(
+    private fun addGene(
         context: CommandContext<CommandSourceStack>,
         geneString: String,
         entities: MutableCollection<out Entity>? = null
@@ -105,12 +106,12 @@ object RemoveGeneCommand {
         val gene = GeneRegistry.fromIdPath(context.source.registryAccess(), geneString)
             ?: throw IllegalArgumentException("Gene with id $geneString does not exist!")
 
-        return removeGene(context, gene, entities)
+        return addGene(context, gene, entities)
     }
 
-    private fun removeGene(
+    private fun addGene(
         context: CommandContext<CommandSourceStack>,
-        geneToRemove: Holder<Gene>,
+        geneToAdd: Holder<Gene>,
         entities: MutableCollection<out Entity>? = null
     ): Int {
 
@@ -118,9 +119,9 @@ object RemoveGeneCommand {
             entities?.mapNotNull { it as? LivingEntity } ?: listOfNotNull(context.source.entity as? LivingEntity)
 
         if (targets.size == 1) {
-            handleSingleTarget(context, targets.first(), geneToRemove)
+            handleSingleTarget(context, targets.first(), geneToAdd)
         } else {
-            handleMultipleTargets(context, targets, geneToRemove)
+            handleMultipleTargets(context, targets, geneToAdd)
         }
 
         return 1
@@ -129,24 +130,23 @@ object RemoveGeneCommand {
     private fun handleSingleTarget(
         context: CommandContext<CommandSourceStack>,
         target: LivingEntity,
-        geneHolder: Holder<Gene>
+        geneHolder: Holder<Gene>,
     ) {
+        val geneWasAdded = addGeneToTarget(target, geneHolder)
 
-        val success = removeGeneFromTarget(target, geneHolder)
-
-        if (success) {
+        if (geneWasAdded) {
             val component =
-                ModLanguageProvider.Commands.REMOVE_SINGLE_SUCCESS.toComponent(
+                ModLanguageProvider.Commands.ADD_SINGLE_SUCCESS.toComponent(
                     Gene.getNameComponent(geneHolder),
-                    target.displayName
+                    target.name
                 )
 
             context.source.sendSuccess({ component }, false)
         } else {
             val component =
-                ModLanguageProvider.Commands.REMOVE_SINGLE_FAIL.toComponent(
+                ModLanguageProvider.Commands.ADD_SINGLE_FAIL.toComponent(
                     Gene.getNameComponent(geneHolder),
-                    target.displayName
+                    target.name
                 )
 
             context.source.sendFailure(component)
@@ -160,33 +160,49 @@ object RemoveGeneCommand {
     ) {
         var amountSuccess = 0
         var amountFail = 0
-        for (target in targets) {
-            val success = removeGeneFromTarget(target, geneHolder)
 
-            if (success) amountSuccess++ else amountFail++
+        for (target in targets) {
+            val geneWasAdded = addGeneToTarget(target, geneHolder)
+            if (geneWasAdded) amountSuccess++ else amountFail++
         }
 
         if (amountSuccess != 0) {
             val component =
-                ModLanguageProvider.Commands.REMOVE_MULTIPLE_SUCCESS.toComponent(
+                ModLanguageProvider.Commands.ADD_MULTIPLE_SUCCESS.toComponent(
                     Gene.getNameComponent(geneHolder),
                     amountSuccess
                 )
-            context.source.sendSuccess({ component }, true)
+
+            context.source.sendSuccess({ component }, false)
         }
+
         if (amountFail != 0) {
             val component =
-                ModLanguageProvider.Commands.REMOVE_MULTIPLE_FAIL.toComponent(
+                ModLanguageProvider.Commands.ADD_MULTIPLE_FAIL.toComponent(
                     Gene.getNameComponent(geneHolder),
                     amountFail
                 )
             context.source.sendFailure(component)
         }
-
     }
 
-    private fun removeGeneFromTarget(target: LivingEntity, geneToRemove: Holder<Gene>): Boolean {
-        val success = target.removeGene(geneToRemove)
+    private fun addGeneToTarget(
+        target: LivingEntity,
+        geneHolder: Holder<Gene>,
+    ): Boolean {
+        val alreadyHasGene = target.hasGene(geneHolder)
+        if (alreadyHasGene) {
+            GeneticsResequenced.LOGGER.info("Tried to add gene ${geneHolder.key!!.location()} to ${target.name.string}, but they already have it!")
+            return false
+        }
+
+        if (!geneHolder.value().canEntityHave(target)) {
+            GeneticsResequenced.LOGGER.info("Tried to add gene ${geneHolder.key!!.location()} to ${target.name.string}, but they can't have it!")
+            return false
+        }
+
+        val success = target.addGene(geneHolder)
+
         return success
     }