+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/command/gene/RemoveGeneCommand.kt
@@ -1,13 +1,11 @@
-package dev.aaronhowser.mods.geneticsresequenced.command
+package dev.aaronhowser.mods.geneticsresequenced.command.gene
 
 import com.mojang.brigadier.arguments.StringArgumentType
 import com.mojang.brigadier.builder.ArgumentBuilder
 import com.mojang.brigadier.context.CommandContext
-import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.Gene
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.GeneRegistry
-import dev.aaronhowser.mods.geneticsresequenced.attachment.GenesData.Companion.addGene
-import dev.aaronhowser.mods.geneticsresequenced.attachment.GenesData.Companion.hasGene
+import dev.aaronhowser.mods.geneticsresequenced.attachment.GenesData.Companion.removeGene
 import dev.aaronhowser.mods.geneticsresequenced.command.ModCommands.SUGGEST_GENE_RLS
 import dev.aaronhowser.mods.geneticsresequenced.command.ModCommands.SUGGEST_GENE_STRINGS
 import dev.aaronhowser.mods.geneticsresequenced.datagen.ModLanguageProvider
@@ -21,7 +19,7 @@ import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.entity.Entity
 import net.minecraft.world.entity.LivingEntity
 
-object AddGeneCommand {
+object RemoveGeneCommand {
 
     private const val GENE_RL_ARGUMENT = "geneRl"
     private const val GENE_STRING_ARGUMENT = "geneString"
@@ -29,7 +27,7 @@ object AddGeneCommand {
 
     fun register(): ArgumentBuilder<CommandSourceStack, *> {
         return Commands
-            .literal("addGene")
+            .literal("remove")
             .requires { it.hasPermission(2) }
             .then(
                 Commands
@@ -41,7 +39,7 @@ object AddGeneCommand {
                             .then(
                                 Commands.argument(TARGET_ARGUMENT, EntityArgument.entities())
                                     .executes { cmd ->
-                                        addGene(
+                                        removeGene(
                                             cmd,
                                             StringArgumentType.getString(cmd, GENE_STRING_ARGUMENT),
                                             EntityArgument.getEntities(cmd, TARGET_ARGUMENT)
@@ -49,7 +47,7 @@ object AddGeneCommand {
                                     }
                             )
                             .executes { cmd ->
-                                addGene(
+                                removeGene(
                                     cmd,
                                     StringArgumentType.getString(cmd, GENE_STRING_ARGUMENT),
                                     entities = null
@@ -67,7 +65,7 @@ object AddGeneCommand {
                                 Commands
                                     .argument(TARGET_ARGUMENT, EntityArgument.entities())
                                     .executes { cmd ->
-                                        addGene(
+                                        removeGene(
                                             cmd,
                                             ResourceLocationArgument.getId(cmd, GENE_RL_ARGUMENT),
                                             EntityArgument.getEntities(cmd, TARGET_ARGUMENT)
@@ -75,7 +73,7 @@ object AddGeneCommand {
                                     }
                             )
                             .executes { cmd ->
-                                addGene(
+                                removeGene(
                                     cmd,
                                     ResourceLocationArgument.getId(cmd, GENE_RL_ARGUMENT),
                                     entities = null
@@ -86,18 +84,19 @@ object AddGeneCommand {
 
     }
 
-    private fun addGene(
+    private fun removeGene(
         context: CommandContext<CommandSourceStack>,
         geneRl: ResourceLocation,
         entities: MutableCollection<out Entity>? = null
     ): Int {
+
         val gene = GeneRegistry.fromResourceLocation(context.source.registryAccess(), geneRl)
             ?: throw IllegalArgumentException("Gene with id $geneRl does not exist!")
 
-        return addGene(context, gene, entities)
+        return removeGene(context, gene, entities)
     }
 
-    private fun addGene(
+    private fun removeGene(
         context: CommandContext<CommandSourceStack>,
         geneString: String,
         entities: MutableCollection<out Entity>? = null
@@ -106,12 +105,12 @@ object AddGeneCommand {
         val gene = GeneRegistry.fromIdPath(context.source.registryAccess(), geneString)
             ?: throw IllegalArgumentException("Gene with id $geneString does not exist!")
 
-        return addGene(context, gene, entities)
+        return removeGene(context, gene, entities)
     }
 
-    private fun addGene(
+    private fun removeGene(
         context: CommandContext<CommandSourceStack>,
-        geneToAdd: Holder<Gene>,
+        geneToRemove: Holder<Gene>,
         entities: MutableCollection<out Entity>? = null
     ): Int {
 
@@ -119,9 +118,9 @@ object AddGeneCommand {
             entities?.mapNotNull { it as? LivingEntity } ?: listOfNotNull(context.source.entity as? LivingEntity)
 
         if (targets.size == 1) {
-            handleSingleTarget(context, targets.first(), geneToAdd)
+            handleSingleTarget(context, targets.first(), geneToRemove)
         } else {
-            handleMultipleTargets(context, targets, geneToAdd)
+            handleMultipleTargets(context, targets, geneToRemove)
         }
 
         return 1
@@ -130,23 +129,24 @@ object AddGeneCommand {
     private fun handleSingleTarget(
         context: CommandContext<CommandSourceStack>,
         target: LivingEntity,
-        geneHolder: Holder<Gene>,
+        geneHolder: Holder<Gene>
     ) {
-        val geneWasAdded = addGeneToTarget(target, geneHolder)
 
-        if (geneWasAdded) {
+        val success = removeGeneFromTarget(target, geneHolder)
+
+        if (success) {
             val component =
-                ModLanguageProvider.Commands.ADD_SINGLE_SUCCESS.toComponent(
+                ModLanguageProvider.Commands.REMOVE_SINGLE_SUCCESS.toComponent(
                     Gene.getNameComponent(geneHolder),
-                    target.name
+                    target.displayName
                 )
 
             context.source.sendSuccess({ component }, false)
         } else {
             val component =
-                ModLanguageProvider.Commands.ADD_SINGLE_FAIL.toComponent(
+                ModLanguageProvider.Commands.REMOVE_SINGLE_FAIL.toComponent(
                     Gene.getNameComponent(geneHolder),
-                    target.name
+                    target.displayName
                 )
 
             context.source.sendFailure(component)
@@ -160,49 +160,33 @@ object AddGeneCommand {
     ) {
         var amountSuccess = 0
         var amountFail = 0
-
         for (target in targets) {
-            val geneWasAdded = addGeneToTarget(target, geneHolder)
-            if (geneWasAdded) amountSuccess++ else amountFail++
+            val success = removeGeneFromTarget(target, geneHolder)
+
+            if (success) amountSuccess++ else amountFail++
         }
 
         if (amountSuccess != 0) {
             val component =
-                ModLanguageProvider.Commands.ADD_MULTIPLE_SUCCESS.toComponent(
+                ModLanguageProvider.Commands.REMOVE_MULTIPLE_SUCCESS.toComponent(
                     Gene.getNameComponent(geneHolder),
                     amountSuccess
                 )
-
-            context.source.sendSuccess({ component }, false)
+            context.source.sendSuccess({ component }, true)
         }
-
         if (amountFail != 0) {
             val component =
-                ModLanguageProvider.Commands.ADD_MULTIPLE_FAIL.toComponent(
+                ModLanguageProvider.Commands.REMOVE_MULTIPLE_FAIL.toComponent(
                     Gene.getNameComponent(geneHolder),
                     amountFail
                 )
             context.source.sendFailure(component)
         }
+
     }
 
-    private fun addGeneToTarget(
-        target: LivingEntity,
-        geneHolder: Holder<Gene>,
-    ): Boolean {
-        val alreadyHasGene = target.hasGene(geneHolder)
-        if (alreadyHasGene) {
-            GeneticsResequenced.LOGGER.info("Tried to add gene ${geneHolder.key!!.location()} to ${target.name.string}, but they already have it!")
-            return false
-        }
-
-        if (!geneHolder.value().canEntityHave(target)) {
-            GeneticsResequenced.LOGGER.info("Tried to add gene ${geneHolder.key!!.location()} to ${target.name.string}, but they can't have it!")
-            return false
-        }
-
-        val success = target.addGene(geneHolder)
-
+    private fun removeGeneFromTarget(target: LivingEntity, geneToRemove: Holder<Gene>): Boolean {
+        val success = target.removeGene(geneToRemove)
         return success
     }