+++ CHANGELOG.md
@@ -234,3 +234,4 @@
     - `#geneticsresequenced:fireball` is now `#geneticsresequenced:activates_shoot_fireball_gene`
     - `#geneticsresequenced:magnet_blacklist` is now `#geneticsresequenced:item_magnet_gene_blacklist`
 - Removed the tag `#geneticsresequenced:wooly`, it just uses `#c:tools/shear` now
+- Added EMI tag translations